Basic Menu Button Functionality & StatManager
So we hooked up buttons to simply move between scenes no fancy animations or loading screens yet.

StatManager stars in the TitleScreen so its always active, and it currently updates not everything listed, but the main ones within the script just look.

Dmg Popup text is also a thing now, was shifting itself below the ground over time but maybe we fixed that just now hopefully