EnemySelectionUI + ServerRack Funtionality
So ServerRack now has its own InfoUI panel and when it places (only on placing) does it add to the total available "storage".

EnemySelectionScreen is barebones but pretty decent so far, pretty easy to swap and replace stuff for all 8 enemy types, we have 2 done already so should be just polishing up and getting functionality for the rest of it.

Focus on getting a proper UI "theme" even if its blue just commit to it hermano.

DOPA DOWN