Improved MainGame Ground Texture
Used the matching ground texture from the title screen but a bit stripped down and only the large grid cubes, got them to line up perfectly with our existing grid system so it feels super snappy and helpful for placement stuff.

Played with the priority layering for a lot of the UI and material stuff since the new plane material existed on priority 1 and so did our background 1x1 floating platforms so had to up the priority for all the build UI, works into the enemy selection screen, haven't tested in combat mode during a wave so that is still left.