Almost Broke + BitMiner Functionality & UI Basically Done
So started off pretty good with Bit Miner getting setup to have a production rate and a capacity to fill up to and the UI is setup now with two different tabs. Spent like an hour arguing with gpt to setup the animation states for the tab but eventually we got it working. Actually feels not awful but needs visual touch ups. But the animation of the panels sliding looks clean af.

Yeah where we almost lost all of that progress, the cost highlighter thingy i tried to attach to our upgrade cost text objects in the bit miner panel and it stripped EVERY TMP object in the scene of the like lower inspector weird overlay material and yeh we had to manually go one by one and remove and copy ./ paste the TMP component values for each object, like over 100 it was not fun be careful of anything related to font materials or save in github before doing that >:( 👺